+++ .gitignore
@@ -7,3 +7,4 @@
 *.sublime*
 Thumbs.db
 vendor
+deleted_files
