+++ Readme.md
@@ -1,7 +1,7 @@
 # Личный проект «TaskForce»
 
 * Студент: [Михаил Лебедев](https://up.htmlacademy.ru/yii/1/user/806805).
-* Наставник: `unknown`.
+* Наставник: `Martin German`.
 
 ---
