+++ _tests-m2t2v1.php
@@ -1,7 +1,6 @@
 <?php
 
 require_once ('vendor/autoload.php');
-use TaskForce\General\Task;
 use TaskForce\General\AvailableActions;
 use TaskForce\General\AddTaskAction;
 use TaskForce\General\CancelAction;
@@ -22,25 +21,36 @@ function d($value){
 // ТЕСТИРОВАНИЕ - примеры данных
 
 $users = [];
+// ($taskId, $taskName, $currentStatus, $endDate, $customerId, $contractorId)
 
-$users[] = [1, 'Ivan', Task::STATUS_NEW, '2019-15-11', 1, 2];
-$users[] = [2, 'Ivan', Task::STATUS_RUNNING, '2019-15-11', 1, 2];
+$users[] = [1, 'task1', AvailableActions::STATUS_NEW, '2019-12-11', 1, 2];
+$users[] = [2, 'task2', AvailableActions::STATUS_RUNNING, '2019-12-11', 1, 2];
+$users[] = [3, 'task3', AvailableActions::STATUS_CANCELED, '2019-12-11', 1, 2];
+$users[] = [4, 'task4', AvailableActions::STATUS_NEW, '2019-12-11', 3, NULL];
+$users[] = [5, 'task25', AvailableActions::STATUS_RUNNING, '2019-12-11', 3, ''];
 
-// ТЕСТИРОВАНИЕ ВЫЗОВ ОБЪЕКТА - вручную для каждого изменения (http://localhost/_tests-M2T1.php) !!! 
+// ТЕСТИРОВАНИЕ ВЫЗОВ ОБЪЕКТА - вручную для каждого изменения (http://localhost/_tests-m2t2v1.php) !!! 
 foreach($users as $key => $user) {
 
-    $task = new Task($user[0], $user[1], $user[2], $user[3], $user[4], $user[5]);
+    $AvailableActions = new AvailableActions($user[0], $user[1], $user[2], $user[3], $user[4], $user[5]);
 
     print('<br> show_current_status: ');
-    echo $task->getCurrentStatus();
-    
-    // Запустить $task = new AvailableAction
-    $availableActions = new AvailableActions($user[0], $user[1], $user[2], $user[3], $user[4], $user[5]);
+    echo $AvailableActions->getCurrentStatus();
 
     print('<br> show Available Actions before');
-    // Запустить $task = new AvailableAction($task->getAvailableActions(2)
-    d($availableActions->getAvailableActions($user[4]));
+    d($AvailableActions->getAvailableActions($user[4])); // $user[4] и $user[5] - это ИД заказчика и исполнителя
 
+    print('<br> show Next Status: ');
+    // Запустить $task = new AvailableAction($task->getAvailableActions(2)
+    echo $AvailableActions->getNextStatus(AvailableActions::ACTION_SET_CONTRACTOR); // по действию находится следующий статус.
+
+    print('<br> <br> show Available Actions after');
+    d($AvailableActions->getAvailableActions($user[4])); // $user[4] и $user[5] - это ИД заказчика и исполнителя соотвественно
+
+    print 'show Statuses: '; d($AvailableActions->getStatuses());
+    print 'show Actions: '; d($AvailableActions->getActions());
+
+    /*
     // Наверное, пока не очень понимаю как может быть нужно 
     $actionInfo = new AddTaskAction;
     d($actionInfo->verifyAccess($availableActions));
@@ -62,6 +72,6 @@ foreach($users as $key => $user) {
 
     $actionInfo = new SetExecutorAction;
     d($actionInfo->verifyAccess($availableActions));
-
+*/
     print('<hr>');
 }
