+++ classes/General/AbstractAction.php
@@ -4,9 +4,9 @@ namespace TaskForce\General;
 
 abstract class AbstractAction
 {
-    abstract public static function getActionId();
+    abstract public static function getActionSymbol();
 
-    abstract public static function verifyAccess(Task $availableActions);
+    abstract public static function verifyAccess(AvailableActions $availableActions);
 
     abstract public static function getActionName();
 }
